# src/docs/swagger.yaml
definitions:
  db.USER:
    properties:
      full_name:
        type: string
      plain_text_password:
        type: string
      user_name:
        type: string
    type: object
  enc.REQUEST:
    properties:
      key:
        type: string
      text:
        type: string
    type: object
  enc.RESPONSE:
    properties:
      result:
        type: string
    type: object
info:
  contact: {}
paths:
  /decrypt:
    post:
      consumes:
      - application/json
      parameters:
      - description: values to decrypt
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/enc.REQUEST'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/enc.RESPONSE'
        "400":
          description: Bad Request Error
          schema:
            type: string
  /encrypt:
    post:
      consumes:
      - application/json
      parameters:
      - description: values to encrypt
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/enc.REQUEST'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/enc.RESPONSE'
  /keys:
    get:
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
  /test:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - example
  /user:
    post:
      consumes:
      - application/json
      parameters:
      - description: values to encrypt
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/db.USER'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.USER'
  /user/{username}:
    get:
      parameters:
      - description: User to get
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.USER'
swagger: "2.0"